Corrected issues with cart controls, adjusted camera, moved everything closer to world origin to prevent issue with physics
Corrected bug with cart controls only trigger when overlapping objects with 'Generate Overlap Events' enabled. Switched it to a trace system, so this is no longer needed and will not be a problem going forward.

Set the default texture pool in project settings from 64mb to 8000mb (allows for ram to be dedicated to textures)

Enabled 'Virtual Texture Support' in project settings, should help with optimizing texture rendering in the future.

Turned off Motion Blur

Fixed bug with camera being able to twist in a way that wasn't relative to the cart

Added a limit for how much the camera can tilt above and below the cart. Also added a function that determines if would be in that range before adding the rotation.

Changed carts frame Linear Damping from 0.01 to 1.5, for more controlled speeds.

Disabled 'CameraMode' Velocity for now.

If player camera is in 'Locked' mode, it will switch to 'Free' while airborne or on their side, then back to 'Locked' when they are back on their wheels again.

Moved everything in the level closer to the world origin, physics was misbehaving the further it was away from it.