Tennis Racket functionality and Slingshot upgrade
There is a bug with the slingshot getting the way of the player's other hand and the projectile will shoot the player/fall to the ground.

Also the weapons lag behind while the player is driving in the cart.

AI shot prediction

New projectile with slingshot

Tennis racket head moves towards the projectile to intercept.